refactor(home-gui): relocate gui related classes into dedicated packages
